pyproject.toml
[tool.poetry]
name = "restr"
version = "0.1.0"
description = ""
authors = ["punitarani <punitsai36@gmail.com>"]
readme = "README.rst"

[tool.poetry.dependencies]
python = "^3.10"
coverage = "^6.4.4"
psutil = "^5.9.2"
pylint = "^2.15.3"
pytest = "^7.1.3"
python-dotenv = "^0.21.0"
selenium = "^4.4.3"
webdriver-manager = "^3.8.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
